+++ Changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - [DCAEGEN2-3427] Refactored PNF event consumption to use Spring KafkaListener instead of SDK MessageRouterSubscriber
 - [DCAEGEN2-3427] Added auto.offset.reset=earliest to prevent missing events due to race condition
+- [DCAEGEN2-3428] Refactored PNF event publishing to use Spring KafkaTemplate instead of dmaap sdk's MessageRouterPublisher
 
 ## [1.11.0] - 2026/01/01